requirements.txt
openai>=1.12.0
google-cloud-bigquery
streamlit
pandas
matplotlib
db-dtypes	


